Change data merge to setblock
Having the block set to air then back to a dropper kicks the player out of the menu in case s/he was inside of it while the custom crafter was being created. This prevents a bug which I have seen many times where the block is renamed to 'Custom Crafter', but you need to go out and back into the dropper to see the renamed block. Until you re-open the menu, it will still just say 'Dropper' as your block entity name.