Add items from the aquatic update
Previously the items were the same as the 1.12 update, with the exception of seagrass. Kelp (7.5%), Sea Pickle(5%), and Scute(2.5%) has been added. Resultantly, some items have reduced chances: Bone(15%), Seagrass(15%), Squid(10%) Additionally, support for SCUBA Gear has been implemented. **IF** the SCUBA Gear module is installed, the chance of waders and leather boots is decreased by 0.5%, and the Tank and Flippers can be acquired this way (each a 0.5% chance), but they have very low durability left (2 and 4 respectively).