Update Wormhole Bottles for 1.14
A change by Mojang in 1.14 resulted in Wormhole Bottles no longer
working if their destination lay in unloaded chunks. This change has
been accommodated for by replacing the `teleport_to_dimension` function
with `set_position` and `set_dimension`, which use the marker area
effect cloud as `@s` instead of the player.

Computational loot tables are now used to check the player's mainhand
and offhand slots for Wormhole Bottles, this avoids the cost of
serializing every player's NBT twice every 16 ticks (which is extremely
expensive).

Due to 1.14 adding the ability to copy NBT between entities, player NBT
is only serialized a single time per Wormhole Bottle in their hands. A
temporary item entity is used to minimize the cost of copying the
stored coordinates into the scoreboard. Once Gamemode 4 provides a
forceloaded chunk, this optimization can be made even better by using a
Jukebox rather than a temporary item entity.

Due to 1.14 adding advanced NBT paths, it is now possible to efficiently
retrieve the coordinates of the offhand Wormhole Bottle. Therefore, this
commit has enabled the ability for the player to consume Wormhole
Bottles held in their offhand and get teleported to the intended
destination.

Proper offhand support has also fixed an exploit which allowed the
player to consume a Wormhole Bottle in their offhand, but in fact get
teleported to the location of a Wormhole Bottle in their mainhand (which
they did not actually consume). Now, they will be correctly teleported
to the location stored in their offhand Wormhole Bottle.