Reassign CustomModelData for Magic in a Bottle
This commit reassigns the CustomModelData for Magic in a Bottle from 2
to 3, to give room for Sunken Treasure to have an advancement icon with
CustomModelData 2. This is required for consistency purposes, as every
other module with both an advancement icon and a custom item using the
same item ID has sequential CustomModelData.