[Submission] Rituals
Rituals is a module that allows you to trigger certain world events using 'recipes' of mobs combined with a block.

### Creating an altar
The Altar is where all rituals take place. The structure can be found in the structure file `rituals:altar`. If you sneak on the redstone block in the centre after building this structure, an altar is spawned.
You can create a 'Water Altar' by filling the stairs adjacent to the redstone block with water. These are referenced later.

### Sacrifices
A sacrifice takes place when the redstone block is on fire and a mob is present at the fire block. The mob is killed with an animation, and added to a 'stock'. This stock is global and can be used at any altar. Every mob's stock is decreased every 30 seconds. The larger the stock, the more taken away every cycle.
Certain items can be added to the altar by having a sacrificed mob holding one. These are referenced later.

### Recipes
Please note - recipes get harder as the ritual's reward increases + whenever a block is mentioned, it has to be placed above the redstone block in the altar
- Weather
   - Thunder (Needs Water Altar)
      - 4 mobs with a trident
      - 2 drowned
      - Dark prismarine block
   - Rain (Needs Water Altar)
      - 2 drowned
      - Prismarine Block/Prismarine Bricks
   - Clear
      - 2 wither skeletons
      - Glass Block

### Advancements
- *God Have Mercy on Your Soul* Create an Altar for rituals
- *Underwater Prayers* Create a Water Aaltar
- *Please the Gods* Sacrifice a mob to the Altar