Fix issues with teleportation
This commit moves the arrival animation to 'set_dimension', and removes
the dedicated 'arrival_animation' function, which fixes the issues with
TNT not spawning.

The teleport particles now display as expected; this was only possible
by making them 'force', but they are limited to '@a[distance=..32]' to
simulate 'normal' as much as possible.

The explosion 'playsound' was removed to avoid duplicate sounds, if
there is an issue with this, it can be easily re-added. I only removed
it because I had copied code that used '@s', but '@s' is not the player
in this context. I did not want to make it use the special '@a' selector
when the TNT itself already made an explosion sound.

Additionally, the chunk is now forcibly loaded by quickly running
'forceload add' and 'forceload remove'; this is not necessary in vanilla
but additional testing is needed to determine if it is needed on Paper.
'forceload query' is used first to avoid deleting an already-forceloaded
chunk.