Update init and init_check
This new version of ZC needs to be re-initialized due to new scoreboard
objectives being used to track both mainhand and offhand wormhole
positions. Therefore, init_check needs to check for 2.. rather than 1..
and init needs to set 2.

This commit also modifies pack.mcmeta to credit Alluet for this update.