Update Wormhole Bottles for 1.14 (#198)
* Update Wormhole Bottles for 1.14

A change by Mojang in 1.14 resulted in Wormhole Bottles no longer
working if their destination lay in unloaded chunks. This change has
been accommodated for by replacing the `teleport_to_dimension` function
with `set_position` and `set_dimension`, which use the marker area
effect cloud as `@s` instead of the player.

Computational loot tables are now used to check the player's mainhand
and offhand slots for Wormhole Bottles, this avoids the cost of
serializing every player's NBT twice every 16 ticks (which is extremely
expensive).

Due to 1.14 adding the ability to copy NBT between entities, player NBT
is only serialized a single time per Wormhole Bottle in their hands. A
temporary item entity is used to minimize the cost of copying the
stored coordinates into the scoreboard. Once Gamemode 4 provides a
forceloaded chunk, this optimization can be made even better by using a
Jukebox rather than a temporary item entity.

Due to 1.14 adding advanced NBT paths, it is now possible to efficiently
retrieve the coordinates of the offhand Wormhole Bottle. Therefore, this
commit has enabled the ability for the player to consume Wormhole
Bottles held in their offhand and get teleported to the intended
destination.

Proper offhand support has also fixed an exploit which allowed the
player to consume a Wormhole Bottle in their offhand, but in fact get
teleported to the location of a Wormhole Bottle in their mainhand (which
they did not actually consume). Now, they will be correctly teleported
to the location stored in their offhand Wormhole Bottle.

* Replace occurences of § with text components

* Correct comment capitalization to fit the standard

* Correct a comment mix-up

* Fix issues with teleportation

This commit moves the arrival animation to 'set_dimension', and removes
the dedicated 'arrival_animation' function, which fixes the issues with
TNT not spawning.

The teleport particles now display as expected; this was only possible
by making them 'force', but they are limited to '@a[distance=..32]' to
simulate 'normal' as much as possible.

The explosion 'playsound' was removed to avoid duplicate sounds, if
there is an issue with this, it can be easily re-added. I only removed
it because I had copied code that used '@s', but '@s' is not the player
in this context. I did not want to make it use the special '@a' selector
when the TNT itself already made an explosion sound.

Additionally, the chunk is now forcibly loaded by quickly running
'forceload add' and 'forceload remove'; this is not necessary in vanilla
but additional testing is needed to determine if it is needed on Paper.
'forceload query' is used first to avoid deleting an already-forceloaded
chunk.

* Add missing end-of-file line breaks

* Use Unix line breaks for brand-new files

* Remove unnecessary forceload chunk loading

There was temporary forceloading left in for testing reasons--this has
been removed as it is not necessary (even on Paper).

* Re-add forceloads

Apparently there *are* issues with loading chunks, despite them not
happening with my testing. This should put a stop to those issues.

* Update init and init_check

This new version of ZC needs to be re-initialized due to new scoreboard
objectives being used to track both mainhand and offhand wormhole
positions. Therefore, init_check needs to check for 2.. rather than 1..
and init needs to set 2.

This commit also modifies pack.mcmeta to credit Alluet for this update.