Bug Fix
The maximum durability for the moneo tool is now stored inside the NBT of the item in the item's tag at "MaxDurability". Since this only gets set once, ever, it's also much better than the last system. Thanks to misode for the idea, and thanks to Kroppeb for pointing out the last system's flaw.