Add Liquid Minecarts (#221)
* Add Liquid Minecarts

Liquid Minecarts first draft. I threw this together in a few days so I'm sure there are errors or places where performance can be improved.

Video: https://www.youtube.com/watch?v=qQctcD3WvWw

Liquid Minecarts are used to move liquids between tanks. This module needs feedback and testing. The goal in its design is to support any new liquids automatically. Instead of hard-coded values, liquid minecarts copy the scores, tags and textures from the tank they load from.

Minecarts can hold 40% of the capacity of the tank from which they got their liquid from (usually 120 bottles).
Minecarts are created by placing down a hopper minecart and entering the same recipe as a liquid tank.
Minecarts will fill if there is a liquid tank pointed at them and they are empty or the liquid in the tank matches their own.
Minecarts will empty their contents into an empty tank or a tank with matching liquid if they are on top of the tank.
Minecarts will empty completely if they drive over a powered activator rail. This is so automated systems can assure tanks have no residual contents.
Minecarts have no GUI for survival players and have no functionality outside of being able to interface with tanks. Think of them as pipes.

A note on why liquid minecarts are command_block_minecart entities. Originally I used regular minecarts but this caused the passenger tank armour stand to be ejected when going over an activator rail. Minecarts themselves don't support skulls as a custom display block. I didn't want the minecarts to have a GUI because they have no GUI functionality and seeing a hopper interface on a mobile liquid tank will cause players to expect it to be able to bottle and unbottle. Furthermore a GUI will cause the tank to suck items out of hoppers or place them in, potentially messing up auto-filler systems if it picks something up while travelling down a track. Furnace minecarts drive weird as we all know so command block minecart it is! There's an added advantage that the "drain" feature is entered as a command in the minecart itself meaning there's no 20hz clock needed to check for activator rails. As far as I can tell, in survival it acts like a normal minecart. You can't view the command GUI and breaking it doesn't drop a command block.

* Removed leftover functions

* Minor comment position tweaks

* Update gm4_liquid_minecarts/data/liquid_minecarts/functions/load_check.mcfunction

Co-Authored-By: Kroppeb <robbepincket@live.be>

* Update gm4_liquid_minecarts/data/liquid_minecarts/functions/load_check.mcfunction

Co-Authored-By: Kroppeb <robbepincket@live.be>