Moneo Shamir (#259)
* Moneo Shamir

A datapack expansion to metallurgy adding a new shamir, Moneo.

For the shamir to function, #metallurgy_expansion:main must be run in the metallurgy:main function before metallurgy:has_shamir.

* Fixed Structure

After speaking with blue, a solution arose which now makes it straight up compatible with metallurgy.

* Module Name and Translation Key Fix

As discussed, the module name and translation keys have been changed to be more suitable.

* Requested Changes and Optimisations

As complying with the requested changes, the shamir band has been changed to Aluminium to better conform with the idea of supression. Gold tools have also been added. Also fixed a minor issue where the module would reinstall itself every reload. As by Kroppeb's and Misode's notice, the checks have been optimised to reduce the amount of NBT checks.

* Bug Fix

The maximum durability for the moneo tool is now stored inside the NBT of the item in the item's tag at "MaxDurability". Since this only gets set once, ever, it's also much better than the last system. Thanks to misode for the idea, and thanks to Kroppeb for pointing out the last system's flaw.

* Sound Update

Sounds have been updated

* Detection failure fix

For some reason, the module wasn't doing anything. That's now fixed. Yay.

* Rename module folder to gm4_moneo_shamir

* Add Site Category to pack.mcmeta