Tinkering Compressors (#248)
* Tinkering Compressors

A working module, tweaks and a texture for the compressor are still needed.

* Fixed Repeated Compressing

Compressors / Decompressors would keep compressing / decompressing respectively. This update ensures it only happens once, and then the items must be picked up.

* Credits + Texture Addition

Edits to pack.mcmeta in terms of credit were made, and a new texture was added.

* Updated Create Checking

By a suggestion, the check to see if a tinkering compressor should be created checks that there is a piston above the dropper before applying the tag to the shamir.

* Efficiency and Comment improvements

This commit adds proper comment heads to every function as well as the following minor efficiency improvements:

- compress and decompress checks will no longer check the nbt of the item if the id of the item was already matched

- iron is checked first for both compression and decompression

- moved @e selectors from `tinkering_compressors:mode/compress` and `tinkering_compressors:mode/decompress` to `tinkering_compressors:item`

- removed "Idea By" Array from `pack.mcmeta` as we usually don't list those - unless the person made a functioning mock-up of the module

- set compressing mode to only trigger if the tinkering compressor is set to exactly 64 items to better reflect compressor behavior

* Set Required Modules

`block_compressors` was missing from the `required_modules` array.