Add Soul Glass (#371)
* Add Soul Glass

Adds a new glass type which is created by smelting soul sand in a blast furnace. When placed on top of a beacon, the effects are reversed. 
https://docs.google.com/document/d/1aREQrHd4CkpZM2I9pajyL5oerTAsoTBoRPzObOUMTP4/edit?usp=sharing

Requested by Bloo, created by BPR and Denniss.

* Change Hunger Effect to Poison

- Changes hunger to poison
- Bug fix for false positive advancement

* Make Soul Glass Toggle-able

Soul Glass can be toggled off if the brown glass is pushed away 1 block.
Also adds some optimizations to the functions

* Parity Fixes

-beacon beam is now required for a soul glass beacon to emit effects
-if two primary effects are selected, they will both be level I instead of one being level II