Add files via upload
This module allows the player to apply the enchantment glow to most items in the game (look at function "blacklist_items" for a list of blacklisted items). This can be used by players to draw attention to items they find special and want to show off.

To get a "shiny" item the player must simply throw an item near xp orbs. Slowly the item will consume xp from the nearby orbs and once it has collected 150 points worth of xp, it will become "shiny".