Move Helious into Weighted Armor (#407)
* Move Helious to Weighted Armor

This moves the Helious shamir to Weighted Armor, as it was origially not considered one of the packaged shamirs.

* Fix Spelling

Helious was misspelled in previous versions of 1.13+ Metallurgy. The name Helious is meant to be spelt with an "u", as it means "helium-like". It is not a reference to the sun (helios).

* Add backwards compatibility

Old 'Helios' shamirs can be converted to 'Helious' shamirs simply by re-applying them.

* Exclusive branching in check_item_validity

Makes sure only one of the two band checks runs.

* Only update shamir id of 'helios' shamirs

* Make sure Lore is updated