Add Multiplayer Sleeping
Multiplayer Sleeping datapack for Minecraft 1.16

Allows for players to pass night and clear thunder storms when 25% of online players are in beds

Accounts only for players in the overworld. Is compatable with gm4_afk_detector to not take AFK players into account when checking total player count

Clears thunder storms at sunrise. Also clears thunderstorms if players are able to sleep during the day