Add Fulcio Shamir (#402)
* Add Fulcio Shamir

New Fulcio Shamir allows players to place blocks in mid-air. Works really nicely with levity!

Metal is Barimium (Modification)
Goes on a Compass

To place blocks in mid-air, the player must sneak and hold the Fulcio Compass in the main or off hand. A structure void will appear 3 blocks in front of the player's eyes (if the block is air) and a particle will display where the hitbox is. Players can then right click on the structure void and the block will be replace the structure void.

Notes: 
The structure void can be broken in survival, but will not drop anything and will not break the module. 
This module does not mess with other possible structure voids in the world because a "fulcio block" can only be placed in air, and is marked with an AEC which handles setting the structure void back to air.
Fulcio is latin for "prop up" or "support"

* Change particles to a block

Added Bloo to creators, as he made the particles commands

* Add Wiki Link to Fulcio's mcmeta

Co-authored-by: Bloo <contact.bluefire610@gmail.com>