Make TP Jamming more reliable for higher ping
Changes the teleport back mechanic to track when the player moves after using chorus. This should ensure that the player gets correctly teleported even if their ping.

In order to prevent players from TPing into somewhere to snoop around, the player is given blindness and an invulnerable armor stand is spawned to prevent breaking blocks and opening chests