Simplified CMD preserving method
Went back to a simple `execute unless data entity @s Item.tag.CustomModelData run...`  to preserve or apply the CMD since there wasn't really anything else that couldn't be transferred by the item. Also put the lore back at the end since it's meant to be an attribute, but this will probably change later anyways when we think a bit more about how we want to display lore (the CMD thing is more important anyway).