Add Celaro shamir to suppress Desire Lines
Celaro is a new type of Aluminium Band that can be placed on boots,
included in the Desire Lines module when loaded alongside Metallurgy.
When applied to the player's boots, the Celaro shamir will simply
suppress the effects of Desire Lines, similar to Boots of Ostara.

There is also a fun easter egg as requested by Bloo. See what happens
for yourself by standing in tall flowers in survival mode with Celaro
shamir on your boots.