Tunnel bores model cosmetics (#500)
* Added piston head when placed

* Change orientation when travelling

Made the piston change its orientation according to the direction it's travelling. Uses tags to limit NBT operations.

* Reset rotation back to "up"

Made the piston face upwards again when it runs out of rails or fails to place one

* clearer comment

I think this makes the comment as to when the function is supposed to trigger clearer? Although I think you can technically make that function trigger by quickly breaking the block the piston minecart is about to travel to, but it matters very little.

* Update update_direction.mcfunction

Co-authored-by: hozz8528 <hozz8528@users.noreply.github.com>