Poses Pack: Improve reliability of pose modification
Dennis recently fixed an issue where pose would not be updated properly
due to the pose for Head/Body first being overwritten to 0f, which would
cause the NBT tag to be removed, which makes the NBT path used to copy
the new pose fail.

However, this does not fix the case where the first element is exactly
0f: when it is copied to the entity as part of pose modification, the
whole pose will be deleted in the same way it was with the erroneous
code setting the pose to [0f,0f,0f], in turn making it impossible to
copy the remaining elements as the NBT path fails.

This commit constructs the new pose in storage and then copies it to
the entity all at once to avoid NBT being reset partway which would
lead to same issue as Dennis partially fixed. Theoretically this could
be done for Body/Head only as arms and legs have different defaults,
but it is better to be consistent and safe for the other parts of the
armor stand, so this commit applies the storage change to every pose.