Metallurgy: make shamirs survive being placed
This update transitions most Metallurgy items to lib_player_heads and
registers each shamir with a unique ID so that the shamir will survive
being placed and broken instead of turning into a Mundane Band.

Shamirs in old worlds are not automatically transitioned to the new
system, however a function tag will be called upon item smoosh and
unsmoosh that upgrades the shamir to survive being placed and broken.