Metallurgy & lib_player_heads: preserve band on placed/broken shamirs (#593)
* lib_player_heads: look up by name when possible

Currently lib_player_heads repairs player heads based on their texture,
but this means that two functionally different items that share the same
player head texture cannot be repaired without losing data.

The main module affected by this at the moment is Metallurgy. Although
Metallurgy does not use lib_player_heads yet, it uses the same technique
of attempting to repair the player head based on texture. However, this
causes placed Obsidian Casts to lose their shamir, as shamirs of a given
metal all share the same texture.

This commit allows users of lib_player_heads to register their player
head via its SkullOwner.Name tag instead of the texture tag, allowing
for an arbitrary string that uniquely identifies the custom item instead
of a texture that might be shared between multiple items.

Modules should select a name that uniquely identifies the custom item
and add it to `gm4_player_heads:register` under the path `heads[].name`.
The `value` and `name` tags should not be used at the same time.
Otherwise, `id` and `item` should be set in the same manner as before.

* lib_player_heads: ignore items with a specific tag

The selector in the lib_player_heads main function now excludes any item
that has the `gm4_player_heads:1b` tag. In the future, modules should
add this tag to newly spawned items and their item's `heads[].item`, but
lib_player_heads will not add the tag itself as it would cause repaired
items to have an extra tag that they do not have when freshly crafted or
otherwise summoned into the world (causing inventory stacking issues).

* Metallurgy: make shamirs survive being placed

This update transitions most Metallurgy items to lib_player_heads and
registers each shamir with a unique ID so that the shamir will survive
being placed and broken instead of turning into a Mundane Band.

Shamirs in old worlds are not automatically transitioned to the new
system, however a function tag will be called upon item smoosh and
unsmoosh that upgrades the shamir to survive being placed and broken.

* Metallurgy: properly copy skull name on smoosh

Previously the SkullOwner.Name tag was not copied if present, despite
the fact that "legacy migration" was performed when *not* present.
Obviously the intended behavior was to copy it if possible, so this
commit correctly implements the intended behavior.

* lib_player_heads: schedule clear v1.0

lib_player_heads v1.0 forgot to /schedule clear itself if it failed to
load, so to avoid v1.0 and v1.1 ticking at the same time, v1.1 will
clear v1.0's scheduled function.

* lib_player_heads: use display to identify skull

Now instead of using a custom tag `gm4_player_head`, lib_player_heads
will exclude items based on the `display` tag as custom skulls have
unique names and lore anyway. This allows Metallurgy to use the library
to turn broken old-format bands into Mundane Bands instead of having
in-house logic for that, without breaking old-format bands that are just
thrown on the ground directly.

Co-authored-by: Bloo <contact.bluefire610@gmail.com>