particles pack update draft
This is just a draft, so I know parts of the code could be improved or removed, but the idea is that you can make any particle instead of just one from a predefined list.

In a written book:
Page 1: `particle`
Page 2: `<particle name>
Page 3: `small` or `big`

`<particle name>` can be any particle name in the game. For particles that require extra inputs, add it on the same page.
Examples: `dust 1 1 1 1`, `block diamond_block` , `item diamond`

Page 3 is to set the size of the particle cloud, where it defaults to a size of 1, `small` sets the size to 0 and `big` sets the size to 3.