Prefix CustomModelData
Adds prefix "342" to every CustomModelData, followed by a 4-digit identifier.

Using a prefix will allow easier compatibility with other data packs, so resource packs can be merged without requiring a change in the data packs. 342 was chosen because it's the first 3 digits of our language translations. Note that the resource pack will need to be updated alongside this.