Add Animi (#683)
* Implement Animi
Animi is a shamir for metallurgy that makes items soulbound.
This means the items are not dropped on death and are instead respawned with the player.
Animi is the first shamir that can be applied to any item.

Particles and sound accompany this process.
Each items stack with Animi plays one soul particle upon the player's death.
The sound is only played once, regardless of the amount of Animi items.

* Reinforce Spectator Exclusion
Not including this code did not actually cause any issues, as spectators never drop items upon death.
However, this reduces unnecessary code execution.

* Limit Animi compatible items
This limits Animi to tools and aremor by default.
A quaternary search tree is used to effectively scan for an item match.
Other modules may add their own items by use of the function tag `#gm4_animi_shamir:mark_item_validity`.

* Update Comment

* Conditionalize 20Hz Clock
This drastically improves performance by only running the 20Hz clock when there's a dead player waiting to respawn.

* Increase robustness in a synthetical edge-case
Makes sure no items are dropped if there are no items to be dropped.
This edge-case should never occur in normal gameplay.

* Use Spigot-Friendly item detection radius
Spigot spawns items at the players feet upon death, which causes the shamir to fail.
This increases the detection radius to cover both spigot and vanilla.
This comes at the cost of detection precision. It might happen that two player dying at the same place get mixed up items upon respawn.

* Remove duplicate function tree entries

* Add missing tree entries

* Add Carrot and Fungus on a Stick

* Exclude spectators when granting items
This prevents the module from trying to give items to players in specator mode.

* Update Wiki Link
The wiki link in pack.mcmeta previously pointed to Vecto's Wiki page.