Add Dripleaf Filters (#765)
This module adds a new passive functionality to dripleaves. If an item normally lands on a dripleaf, the dripleaf collapses and the item falls through. With this module, the dripleaf will look at the items in a hopper placed in front of it, and if any of the items match the item on the leaf's ID, the item slides off the leaf into the hopper.

Co-authored-by: Bloo <contact.bluefire610@gmail.com>