Add Allay Revitalizing
Allows music to convert vexes into allays. When a player plays a music disc in a jukebox, for 45 seconds the jukebox will add a score to vexes in a 5 block radius. After a vex has listened to 30 seconds of music, it will convert to an allay. If the music disc is taken out or the jukebox is broken before the music time, the jukebox will stop converting vexes.

TO-DO:
- update the particles and sounds for when a vex gets converted
- test it