Shrieking Pneuma
Shrieking is the sonic boom attack from the warden. It is a sneak charge pneuma, but only functions after charging for the max sneak time. Upon using the attack, a beam will be shot up to 15 blocks in the direction the player is looking, dealing 3 hearts to the first living entity it hits