Overflow detection
If the crafting operation doesn't overflow the output barrel, a sound and particle is used to convey that the crafter did something. There are different sounds and particles based on if the craft itself succeeded or failed.

If the crafting operation would overflow the output barrel a sound and particle is used to convey that nothing happened. When this happens, all items stay in the auto crafter buffer (virtual inventory). This means after clearing the output, the piston can be powered again to try again