🚀 0795d7e4e47ad1bda10aeeeb7d16d57456ec86eb Auto crafting (#782)
* Modify CC for Auto Crafting

- add item tag checks
    - add generator for vanilla item tags
- store total id length in a scoreboard
- remove two unused function tags

* Bump CC version to 2.1

New minor version to include item tag checking and total id length

* Add initial auto crafting module

The generator will be in a later commit

* Finish bumping the cc version to 2.1

Missed this...

* Add 1.19 vanilla item tag checks for CC

* Add all 1.19 vanilla recipes to CC

ran the generator
- The generator will be updated to use Beet in the future....

* Add auto crafter block

- auto crafter can be crafted in a CC
- auto crafter can be placed and broken with lib_machines
- auto crafter checks for the multiblock
TO-DO:
- auto crafter functionality

* Implement auto crafting functionality

- auto crafters now work with all CC recipes
--
TODO: add overflow detection

* Return items from failed crafts

if the recipe fails or if there are more items in the buffer than the recipe is set for (e.g. fill the shape, input items, then modify the shape), then the extra items get placed in the output barrel

- also modified the function that drops the buffer items to look better
--
TODO:
overflow detection

* Overflow detection

If the crafting operation doesn't overflow the output barrel, a sound and particle is used to convey that the crafter did something. There are different sounds and particles based on if the craft itself succeeded or failed.

If the crafting operation would overflow the output barrel a sound and particle is used to convey that nothing happened. When this happens, all items stay in the auto crafter buffer (virtual inventory). This means after clearing the output, the piston can be powered again to try again

* Force Top-Left Align Recipes

Mojank decided to make the spyglass recipe inconsistent

* Update generator for mirrored recipes

generator will be ran in next commit

* Run generator for mirrored recipes

* move vanilla recipes to subfolder

- previously every time the generator was run, the auto crafter crafting loot table was deleted
- vanilla recipes are now under a subfolder which gets deleted now instead of the entire crafting folder