Orb of Ankou 1.19 (#780)
* Reduce the time required for sneak charge pneumas

* Fix OoA Sneak Charge Bug

- some sneak charge pneumas didn't properly set the score to 0 if the sneak number was too low

* Rebalance Sneak Charge

- lower duration of rushing and bounding sneak charge effects
- slightly increase the sound of each sneak charge bell sound
- fix draining inconsistency

* Shrieking Pneuma

Shrieking is the sonic boom attack from the warden. It is a sneak charge pneuma, but only functions after charging for the max sneak time. Upon using the attack, a beam will be shot up to 15 blocks in the direction the player is looking, dealing 3 hearts to the first living entity it hits

* 1.19 soul essences

- give new 1.19 mobs soul essence drops
- update some aquatic mobs that didn't have bubbly

* Buff shrieking

Shrieking will now deal 6 hearts to players and mobs (previously 3)

Co-authored-by: Levertion <26185209+Levertion@users.noreply.github.com>