Fix Invalid Minecart Items
Currently hoppers and hopper minecarts can pick up items dropped when machine carts are killed. This happens most commonly with standard minecart unloaders which are designed to pick up the minecart immediately when it's killed.

This PR fixes the issue by scanning hoppers and hopper minecarts for invalid items if no item is dropped when the destroy function is called.