Liquid Tanks ink-bottles on 1.20 signs
Allows ink bottles and glow ink bottles to be used on both the front or back of signs, and adds support for hanging signs. Note that there are no methods by which to prevent the "sign edit" GUI window from opening, so that does sadly still occur when using bottled ink.