🚀 060a5a78bd97584023cd684c14d439d31862cf53 Add Double Doors (#890)
This module allows players to open double doors with ease! Instead of having to click twice, clicking one of the doors will automatically open the other!
Additionally, this module also allows trapdoor placed above doors to open and close with the door! For this to work, the trapdoor must be a bottom half trapdoor, open, and its orientation must match the orientation of the door in at least one of the two positions.

Co-authored-by: SpecialBuilder <specialbuilder32@gmail.com>