Double Doors: Close doors automatically after a player walks through (#920)
Introduces auto-closing to Double Doors. When a player opens a door a marker is spawned which lingers around for a few seconds. If a player walks through that marker within that time, the associated door will be closed (i.e. returned to the state it was in before the player clicked it) automatically behind the player. As with the current Double Doors, sneaking prevents this behavior.
---------
Co-authored-by: SpecialBuilder <specialbuilder32@gmail.com>