🚀 8d5c8e0554a50d8662bf2f0d5374866e3abb749e Allow lib_player_heads to Create Unstackable Player Heads (#937)
Allows `lib_player_heads` to source restore data from loot tables directly instead of fixed a fixed nbt compound in the heads registry. This change removes the need to specify the item nbt twice (once in the heads registry and once in a loot table/function/etc which creates the item within a module) and allows for `lib_player_heads` to create unstackable skulls via loot tables, as this was previously impossible due to the static nature of the raw nbt compound in the heads registry.

---------

Co-authored-by: BluePsychoRanger <mcbluepsychoranger@gmail.com>