Lively Lily Pads - Module (#1139)
Place Decorations on Lily Pads!

This Module is a port of one of my older datapacks, like Blossoming Pots.
I've created an upgrade path, tested it and it works fine.

Decorations include:
- Coral Fans
- Candles
- Cactus Flowers
- Lanterns
- All Torches
- Spore Blossoms

Light emitting decorations use light blocks.
Candles need to be lit with a fire charge or flint and steel.