Use item_model instead of custom_model_data
This will make all items appear as missing textures. It's possible to make them look like leather armor without the RP, but making the item sprites missing textures should "force" the RP to be required. This is important because without the RP the armor is invisible, so we should highly discourage that by making the item textures missing without the RP. I.e. make it basically unplayable without the RP so people don't get invisible armor